# backend/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/cvs/{cvId}/sections": {
            "get": {
                "description": "Retrieve all sections for a specific CV, sorted by order index.",
                "operationId": "SectionsController_list",
                "parameters": [
                    {
                        "name": "cvId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sections retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CvSectionEntity"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "You do not own this CV"
                    },
                    "404": {
                        "description": "CV not found or not owned by user"
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "summary": "List sections in a CV",
                "tags": [
                    "Sections"
                ]
            },
            "post": {
                "description": "Create a new section within a CV. Order index is auto-assigned.",
                "operationId": "SectionsController_create",
                "parameters": [
                    {
                        "name": "cvId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSectionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Section created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CvSectionEntity"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body"
                    },
                    "403": {
                        "description": "You do not own this CV"
                    },
                    "404": {
                        "description": "CV not found or not owned by user"
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "summary": "Create a section",
                "tags": [
                    "Sections"
                ]
            }
        },
        "/api/v1/cvs/{cvId}/sections/{id}": {
            "get": {
                "description": "Retrieve a specific section by its ID.",
                "operationId": "SectionsController_findOne",
                "parameters": [
                    {
                        "name": "cvId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Section retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CvSectionEntity"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "You do not own this CV"
                    },
                    "404": {
                        "description": "Section or CV not found"
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "summary": "Get a section",
                "tags": [
                    "Sections"
                ]
            },
            "patch": {
                "description": "Update the content of a section.",
                "operationId": "SectionsController_update",
                "parameters": [
                    {
                        "name": "cvId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSectionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Section updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CvSectionEntity"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body"
                    },
                    "403": {
                        "description": "You do not own this CV"
                    },
                    "404": {
                        "description": "Section or CV not found"
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "summary": "Update a section",
                "tags": [
                    "Sections"
                ]
            },
            "delete": {
                "description": "Delete a section. Subsequent sections are reindexed automatically.",
                "operationId": "SectionsController_remove",
                "parameters": [
                    {
                        "name": "cvId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "403": {
                        "description": "You do not own this CV"
                    },
                    "404": {
                        "description": "Section or CV not found"
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "summary": "Delete a section",
                "tags": [
                    "Sections"
                ]
            }
        },
        "/api/v1/cvs/{cvId}/sections/{id}/move": {
            "post": {
                "description": "Move a section up or down within the CV by swapping with its neighbor.",
                "operationId": "SectionsController_move",
                "parameters": [
                    {
                        "name": "cvId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MoveSectionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Section moved successfully; returns both swapped sections",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CvSectionEntity"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Cannot move (already first/last section) or invalid direction"
                    },
                    "403": {
                        "description": "You do not own this CV"
                    },
                    "404": {
                        "description": "Section or CV not found"
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "summary": "Move a section",
                "tags": [
                    "Sections"
                ]
            }
        }
    },
    "info": {
        "title": "CV Builder API",
        "description": "API documentation for authentication and user profile management.",
        "version": "1.0.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "access-token": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http",
                "description": "JWT access token sent in Authorization header."
            },
            "refresh-token-cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "refreshToken",
                "description": "Refresh token cookie generated by login/register endpoints."
            }
        },
        "schemas": {
...
            "CvSectionEntity": {
                "type": "object",
                "properties": {}
            },
            "CreateSectionDto": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "PERSONAL_INFO",
                        "enum": [
                            "PERSONAL_INFO",
                            "EXPERIENCE",
                            "EDUCATION",
                            "SKILLS",
                            "PROJECTS",
                            "LANGUAGES",
                            "CERTIFICATIONS",
                            "CUSTOM"
                        ]
                    },
                    "content": {
                        "type": "object",
                        "example": {
                            "fullName": "John Doe",
                            "headline": "Software Engineer"
                        }
                    }
                },
                "required": [
                    "type",
                    "content"
                ]
            },
            "UpdateSectionDto": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "object",
                        "example": {
                            "fullName": "Jane Doe",
                            "headline": "Senior Engineer"
                        }
                    }
                },
                "required": [
                    "content"
                ]
            },
            "MoveSectionDto": {
                "type": "object",
                "properties": {
                    "direction": {
                        "type": "string",
                        "example": "up",
                        "enum": [
                            "up",
                            "down"
                        ]
                    }
                },
                "required": [
                    "direction"
                ]
            }
        }
    }
}